:recyle: Refs. #235 -- refactor json logic trigger validation into reusable validators
